Replace old parser completely by deleting it
