+++ javalin/src/main/java/io/javalin/http/util/RedirectToLowercasePathPlugin.kt
@@ -7,7 +7,7 @@
 package io.javalin.http.util
 
 import io.javalin.Javalin
-import io.javalin.core.PathSegment2
+import io.javalin.core.PathSegment
 import io.javalin.core.createPathParser
 import io.javalin.core.plugin.Plugin
 import io.javalin.core.plugin.PluginLifecycleInit
@@ -26,13 +26,13 @@ class RedirectToLowercasePathPlugin : Plugin, PluginLifecycleInit {
         app.events { e ->
             e.handlerAdded { h ->
                 val parser = createPathParser(h.path, app._conf.ignoreTrailingSlashes)
-                parser.segments.filterIsInstance<PathSegment2.Normal>().map { it.asRegexString() }.forEach {
+                parser.segments.filterIsInstance<PathSegment.Normal>().map { it.asRegexString() }.forEach {
                     if (it != it.toLowerCase()) throw IllegalArgumentException("Paths must be lowercase when using RedirectToLowercasePathPlugin")
                 }
                 parser.segments
-                        .filterIsInstance<PathSegment2.MultipleSegments>()
+                        .filterIsInstance<PathSegment.MultipleSegments>()
                         .flatMap { it.innerSegments }
-                        .filterIsInstance<PathSegment2.Normal>()
+                        .filterIsInstance<PathSegment.Normal>()
                         .map { it.asRegexString() }
                         .forEach {
                             if (it != it.toLowerCase()) {
@@ -55,12 +55,12 @@ class RedirectToLowercasePathPlugin : Plugin, PluginLifecycleInit {
                 val clientSegments = requestUri.split("/").filter { it.isNotEmpty() }.toTypedArray()
                 val serverSegments = createPathParser(entry.path, app._conf.ignoreTrailingSlashes).segments
                 serverSegments.forEachIndexed { i, serverSegment ->
-                    if (serverSegment is PathSegment2.Normal) {
+                    if (serverSegment is PathSegment.Normal) {
                         clientSegments[i] = clientSegments[i].toLowerCase() // this is also a "Normal" segment
                     }
-                    if (serverSegment is PathSegment2.MultipleSegments) {
+                    if (serverSegment is PathSegment.MultipleSegments) {
                         serverSegments.forEach { innerServerSegment ->
-                            if (innerServerSegment is PathSegment2.Normal) {
+                            if (innerServerSegment is PathSegment.Normal) {
                                 // replace the non lowercased part of the segment with the lowercased version
                                 clientSegments[i] = clientSegments[i].replace(innerServerSegment.content, innerServerSegment.content.toLowerCase(), ignoreCase = true)
                             }
